python: python sucks.
Graphics require mesa. mesa requires meson. meson requires python.

During meson's build, a call to import _ctypes is made. ctypes are a
foreign function library (hence their dependency on libffi). Why might
one need those? Why, for opening DLLs or shared libraries of course!

But we have neither. So why is such a call being made?
Best I can tell, meson is either under the impression we are
cross-compiling, or it is painfully confused about the structure of our
system.

Arguably, the simplest solution to get to mesa is simply to build a
shared python and see what happens with meson then. Which is what I'm
going to end up doing, probably. Because after hours and hours of
scouring, searching, and reading, I've come up with basically nothing. I
have no idea *why* meson is doing an import, which has an entire
dependency chain of imports that terminate at _ctypes. All I know is, if
python is built *without* them, we fail because of a missing module. If
python is built *with* them, we fail because dynamic loading is not
supported.

So for now, this is the best stab at python we're going to be getting. I
think.